+++ src/theme-builder/theme-builder.ts
@@ -10,6 +10,14 @@ export const exportTheme = async (theme: Theme) => {
 	const themeColors = new Map(Object.entries(theme.colors));
 
 	const parseColor = (token: string) => {
+		if (token === 'transparent') {
+			return '#0000';
+		}
+
+		if (token.includes('#')) {
+			return token;
+		}
+
 		if (token.indexOf('/') < 0) {
 			return themeColors.get(token) || '#ff0000';
 		}
@@ -17,11 +25,11 @@ export const exportTheme = async (theme: Theme) => {
 		if (!tokenPart) return '#ff0';
 		tokenPart = tokenPart.trim();
 
-		alphaPart ||= "100";
+		alphaPart ||= '100';
 		alphaPart = alphaPart.trim();
 
 		let color = themeColors.get(tokenPart) || '#ff0000';
-		color += (Math.round(Number(alphaPart) * 255)).toString(16);
+		color += (Math.round(Number(alphaPart) * 255)).toString(16).padStart(2, "0");
 		return color;
 	};
 
@@ -61,28 +69,28 @@ export const exportThemeImage = async (theme: Theme) => {
 	const browser = await puppeteer.launch();
 	const page = await browser.newPage();
 	await page.setViewport({ width: 600, height: 350, deviceScaleFactor: 4, });
-	const htmlFile = Bun.file("src/theme-builder/theme.html");
+	const htmlFile = Bun.file('src/theme-builder/theme.html');
 
 	let html = await htmlFile.text();
 
 	const scheme: [string, string][] = [
-		["Theme Name", theme.name],
-		["#000000", theme.colors['editor.bg']],
-		["#ffffff", theme.colors['editor.fg']],
-		["#ff0000", theme.colors['code.red']],
-		["#ffff00", theme.colors['code.yellow']],
-		["#00ff00", theme.colors['code.green']],
-		["#00ffff", theme.colors['code.blue']],
-		["#ff00ff", theme.colors['code.pink']],
-		["#aaaaaa", theme.colors['editor.mg']],
-		["#202020", theme.colors['editor.border']],
-		["#303030", theme.colors['editor.gutter']],
-		["#808080", theme.colors['code.punctuation']],
-		["#404040", theme.colors['code.punctuation.less']],
-		["green", theme.colors['code.comment']],
-		["{major}", version.split('.')[0]!],
-		["{minor}", version.split('.')[1]!],
-		["{patch}", version.split('.')[2]!],
+		['Theme Name', theme.name],
+		['#000000', theme.colors['editor.bg']],
+		['#ffffff', theme.colors['editor.fg']],
+		['#ff0000', theme.colors['code.red']],
+		['#ffff00', theme.colors['code.yellow']],
+		['#00ff00', theme.colors['code.green']],
+		['#00ffff', theme.colors['code.blue']],
+		['#ff00ff', theme.colors['code.pink']],
+		['#aaaaaa', theme.colors['editor.mg']],
+		['#202020', theme.colors['editor.border']],
+		['#303030', theme.colors['editor.gutter']],
+		['#808080', theme.colors['code.punctuation']],
+		['#404040', theme.colors['code.punctuation.less']],
+		['#008000', theme.colors['code.comment']],
+		['{major}', version.split('.')[0]!],
+		['{minor}', version.split('.')[1]!],
+		['{patch}', version.split('.')[2]!],
 	];
 
 	for (const [key, value] of scheme) {